Some try with the modules system. Right now this is still in a very rough phase, as I just wanted to try and see how to do it, with security.
Modules AREN'T plugins, they aren't sandboxed inside a VM, so it should be used with caution, they are dll .net 8 file, that will be run if their SHA256 checksum is added to the CHECKSUM_INTERNAL_MODULES inside EngineModules.cs class!

Will it be kept? Probably, but with a lots of warning to the user if a non-trusted module is added (aka external module, that isn't inside the CHECKSUM_INTERNAL_MODULES).

second commit because I'm an idiot that forgot to add the testModule source file